Заменил application.properties на application.yaml В сущности поменял поля вес и рост на float

+++ src/main/java/ru/sportsDiary/entity/Sportsman.java
@@ -31,9 +31,9 @@ public class Sportsman {
     private Date birthDate;
 
     @Column(name = "height")
-    private Integer height;
+    private Float height;
 
     @Column(name = "weight")
-    private Integer weight;
+    private Float weight;
 
 }
